rpmvercmp.py: tool compares two rpm versions
rpmvercmp.py compares two rpm epoch:version-release strings and return
a result. The advantage is it use the versionCompare function from rpm,
so no yet another third-party reimplementation of such algorithm is
included. And an another one - it works on strings, so no rpm packages
needs to be created for purpose of this tool.